Conflicts:
	editor/editor-api/source/jetbrains/mps/openapi/editor/cells/EditorCell.java
	editor/editor-runtime/source/jetbrains/mps/nodeEditor/EditorComponent.java
	editor/editor-runtime/source/jetbrains/mps/nodeEditor/Highlighter.java
	editor/editor-runtime/source/jetbrains/mps/nodeEditor/cells/EditorCell_Basic.java
	languages/baseLanguage/baseLanguage/languageModels/editor.mps
	languages/languageDesign/editor/source_gen/jetbrains/mps/lang/editor/scripts/EditorContext_MigrationScript.java
	plugins/mpsdevkit/solution/devkit/actions.mps
	plugins/mpsjava/platform/models/jetbrains/mps/ide/java/platform/highlighters.mps
	plugins/mpsjava/platform/models/jetbrains/mps/ide/java/platform/refactorings.mps
	plugins/vcs/common/models/changesmanager/editor.mps
	plugins/vcs/common/models/diff/ui/common.mps
	plugins/vcs/common/source_gen/jetbrains/mps/vcs/changesmanager/editor/BaseVersionEditorComponent.java
	workbench/mps-editor/solution/actions.mps
	workbench/mps-workbench/models/actions.mps

+++ core/kernel/source/jetbrains/mps/smodel/behaviour/BehaviorReflection.java
@@ -34,8 +34,8 @@ public class BehaviorReflection {
     OUR_DEFAULT_VALUE.put(Short.class, (short) 0);
     OUR_DEFAULT_VALUE.put(short.class, (short) 0);
 
-    OUR_DEFAULT_VALUE.put(Integer.class, (int) 0);
-    OUR_DEFAULT_VALUE.put(int.class, (int) 0);
+    OUR_DEFAULT_VALUE.put(Integer.class, 0);
+    OUR_DEFAULT_VALUE.put(int.class, 0);
 
     OUR_DEFAULT_VALUE.put(Long.class, (long) 0);
     OUR_DEFAULT_VALUE.put(long.class, (long) 0);
